refactor: remove components question
